refactor: Make adjustment for code readability
