+++ src/repositories/hooks/useRepositoriesFetch.ts
@@ -2,7 +2,6 @@ import { useContext, useEffect, useState } from "react";
 import { AppContext } from "../../app/AppContext";
 import { useLazyQuery } from "@apollo/client";
 import { SEARCH_REPOSITORIES } from "../queries/searchRepositories";
-import { useNotifications } from "@toolpad/core";
 import { Repository } from "../../shared/types/repository";
 import { mapQueryResponse } from "../utils/dataMapper";
 import { Pagination } from "../../shared/types/graphql";
@@ -40,7 +39,6 @@ export const useRepositoriesFetch = () => {
       startCursor: "",
     },
   });
-  const notifications = useNotifications();
   const { searchQuery } = useContext(AppContext);
   const [getRepositories, { loading, error, data }] =
     useLazyQuery(SEARCH_REPOSITORIES);
@@ -75,21 +73,13 @@ export const useRepositoriesFetch = () => {
   }, [queryParameters, getRepositories]);
 
   useEffect(() => {
-    // Inform about data fetching issue
-    if (!loading && error) {
-      notifications.show("There was an error while fetching data", {
-        severity: "error",
-        autoHideDuration: 3000,
-      });
-    }
-
     if (!loading && !error && data) {
       setResults({
         data: mapQueryResponse(data),
         pagination: data.search.pageInfo,
       });
     }
-  }, [loading, error, data, notifications]);
+  }, [loading, error, data]);
 
   const changePage = (page: "prev" | "next") => {
     if (page === "prev") {
@@ -114,6 +104,7 @@ export const useRepositoriesFetch = () => {
   };
 
   return {
+    error,
     loading,
     results,
     changePage,
